extract SelectionStore
