+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Selection/Selection.js
@@ -4,7 +4,7 @@ import {action, autorun, observable} from 'mobx';
 import type {IObservableValue} from 'mobx'; // eslint-disable-line import/named
 import {observer} from 'mobx-react';
 import {MultiItemSelection} from '../../components';
-import SelectionStore from './stores/SelectionStore';
+import SelectionStore from '../../stores/SelectionStore';
 import DatagridOverlay from './DatagridOverlay';
 import selectionStyles from './selection.scss';
