+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/SelectionStore/index.js
@@ -1,0 +1,4 @@
+// @flow
+import SelectionStore from './SelectionStore';
+
+export default SelectionStore;
